.dagger/diff_parser/requirements.txt
gitpython>=3.1.0
pydantic>=2.0.0
pytest>=7.0.0
PyGithub>=2.0.0